PythonChatAppCode-main/PythonChatAppCode-main/requirements.txt
flask
flask-cors
firebase-admin
google-generativeai
gunicorn
